Complete integration of Prompt and Editor fragments.
Just found out how to properly format commit messages :). The
modifications list is long and was not kept track of, will be a lot
more thorough from now on.

Big focus of this commit was modularization of various components, for
more scalable code. Directory fragment is fully integrated with the new
fragments, starting work on Web fragment.